oldapps/requirements.txt
pandas==2.2.3
Pillow==10.0.0
python-dotenv==1.1.1
Requests==2.32.3
streamlit==1.49.1
streamlit_option_menu==0.4.0
